requirements.txt
# Core API framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# PDF processing
pypdf==5.1.0
pdfplumber==0.11.4

# Embeddings and similarity search
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Development and testing
pytest==8.3.4
httpx==0.28.1
